# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Keeper API Documentation",
        "contact": {
            "email": "adityanaidu.work@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get list of tasks",
                "description": "Returns list of tasks with filters applied",
                "operationId": "getTasks",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "New",
                                "Incomplete",
                                "Complete"
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Filter by due date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by priority",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "High",
                                "Low",
                                "Medium"
                            ]
                        }
                    },
                    {
                        "name": "note_content",
                        "in": "query",
                        "description": "Filter by note content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "note_created_at",
                        "in": "query",
                        "description": "Filter by note creation date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of tasks per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order of the results",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Order by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task with multiple notes and attachments",
                "description": "Creates a task with multiple notes, each having multiple attachments.",
                "operationId": "createTaskWithNotesAndAttachments",
                "requestBody": {
                    "description": "Task data with notes and attachments",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subject": {
                                        "description": "Subject of the task",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "Start date of the task (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "description": "Due date of the task (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "description": "Status of the task",
                                        "type": "string"
                                    },
                                    "priority": {
                                        "description": "Priority of the task",
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Note"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Note": {
                "properties": {
                    "subject": {
                        "description": "Subject of the note",
                        "type": "string"
                    },
                    "note": {
                        "description": "Content of the note",
                        "type": "string"
                    },
                    "attachments": {
                        "type": "array",
                        "items": {
                            "description": "Attachment file",
                            "type": "string",
                            "format": "binary"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}